server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
tenseal
pyyaml
pydantic
torch>=2.1.0
transformers>=4.33.0
